requirements.txt
# Django Core
Django==5.2.7
asgiref==3.7.2
sqlparse==0.4.4

# Database
psycopg2-binary==2.9.7
django-environ==0.10.0

# Forms & UI
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.0.1
django-humanize==0.4.1

# Celery & Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-redis==5.3.0

# Production & Deployment
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.40.0
boto3==1.28.62
django-storages==1.13.2

# Security & Cryptography
cryptography==41.0.7
python-decouple==3.8

# Utilities
requests==2.31.0
chardet==5.2.0
python-magic==0.4.27
django-q2==1.4.1
sqlalchemy==2.0.23

# OCR & Image Processing
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3

# Development (optional)
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.14.0